83. alterações

+++ app/controllers/HomeController.scala
@@ -78,21 +78,21 @@ extends AbstractController(cc) with play.api.i18n.I18nSupport {
     db.withConnection { conn =>
       val ps = conn.prepareStatement("select cd_livro from tb_livro where cd_livro = ?")
       ps.setInt(1,cdLiv)
-      ps.execute()
+      val res = ps.executeQuery()
       val ps2 = conn.prepareStatement("select cd_cliente from tb_cliente where cd_cliente = ?")
       ps2.setInt(1,cdCli)
-      ps2.execute()
-      if(ps != null && ps2 != null){
+      val res2 = ps2.executeQuery()
+      if(res.next() || res2.next()){
         val ps3 = conn.prepareStatement("select count(cd_livro) from tb_movimento where cd_livro = ? and cd_cliente = ?")
         ps3.setInt(1,cdLiv)
         ps3.setInt(2,cdCli)
-        val res = ps3.executeQuery()
-        res.next()
+        val res3 = ps3.executeQuery()
+        res3.next()
         val ps4 = conn.prepareStatement("select nr_quantidade from tb_livro where cd_livro = ?")
         ps4.setInt(1,cdLiv)
-        val res2 = ps4.executeQuery()
-        res2.next()
-        if(res2.getInt(1)-res.getInt(1) > 0){
+        val res4 = ps4.executeQuery()
+        res4.next()
+        if(res4.getInt(1) - res3.getInt(1) > 0){
           val ps5 = conn.prepareStatement("insert into tb_movimento(cd_livro,cd_cliente) values (?,?)")
           ps5.setInt(1,cdLiv)
           ps5.setInt(2,cdCli)
@@ -102,20 +102,26 @@ extends AbstractController(cc) with play.api.i18n.I18nSupport {
           ps6.execute()
         }
       }
-      Redirect("/sM")
+      Redirect("/movimentos")
     }
   }
   
   def dev(cdLiv: Int, cdCli: Int) = Action {
     db.withConnection{ conn =>
-      val ps = conn.prepareStatement("delete from tb_movimento where cd_livro=? and cd_cliente = ?")
+      val ps = conn.prepareStatement("select * from tb_movimento where cd_livro = ? and cd_cliente = ?")
       ps.setInt(1,cdLiv)
       ps.setInt(2,cdCli)
-      ps.execute()
-      val ps2 = conn.prepareStatement("update tb_livro set nr_quantidade = nr_quantidade+1 where cd_livro = ?")
-      ps2.setInt(1,cdLiv)
-      ps2.execute()
-      Redirect("/sM")
+      val res = ps.executeQuery()
+      if(res.next()){
+        val ps = conn.prepareStatement("delete from tb_movimento where cd_livro=? and cd_cliente = ?")
+        ps.setInt(1,cdLiv)
+        ps.setInt(2,cdCli)
+        ps.execute()
+        val ps2 = conn.prepareStatement("update tb_livro set nr_quantidade = nr_quantidade+1 where cd_livro = ?")
+        ps2.setInt(1,cdLiv)
+        ps2.execute()
+      }
+      Redirect("/movimentos")
     }
   }
